Move signin and getSession and add favicon
